requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
idna==3.3
idna-ssl==1.1.0
multidict==5.2.0
python-dotenv==0.19.1
tabulate==0.8.9
typing-extensions==3.10.0.2
yarl==1.7.2
